LPS-102731 Rename FileEntry contributors
