+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/info/display/contributor/FileEntryInfoDisplayContributor.java
@@ -46,7 +46,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alejandro Tardín
  */
 @Component(service = InfoDisplayContributor.class)
-public class DLFileEntryInfoDisplayContributor
+public class FileEntryInfoDisplayContributor
 	implements InfoDisplayContributor<FileEntry> {
 
 	@Override
@@ -144,7 +144,7 @@ public class DLFileEntryInfoDisplayContributor
 			return null;
 		}
 
-		return new DLFileEntryInfoDisplayObjectProvider(fileEntry);
+		return new FileEntryInfoDisplayObjectProvider(fileEntry);
 	}
 
 	@Override
