+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/info/display/contributor/field/FileEntryDownloadURLInfoDisplayContributorField.java
@@ -14,8 +14,10 @@
 
 package com.liferay.document.library.web.internal.info.display.contributor.field;
 
+import com.liferay.document.library.util.DLURLHelper;
 import com.liferay.info.display.contributor.field.InfoDisplayContributorField;
 import com.liferay.info.display.contributor.field.InfoDisplayContributorFieldType;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
@@ -23,6 +25,7 @@ import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import java.util.Locale;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Alejandro Tardín
@@ -31,28 +34,37 @@ import org.osgi.service.component.annotations.Component;
 	property = "model.class.name=com.liferay.document.library.kernel.model.DLFileEntry",
 	service = InfoDisplayContributorField.class
 )
-public class DLFileEntryVersionInfoDisplayContributorField
+public class FileEntryDownloadURLInfoDisplayContributorField
 	implements InfoDisplayContributorField<FileEntry> {
 
 	@Override
 	public String getKey() {
-		return "version";
+		return "downloadURL";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
 		return LanguageUtil.get(
-			ResourceBundleUtil.getBundle(locale, getClass()), "version");
+			ResourceBundleUtil.getBundle(locale, getClass()), "download-url");
 	}
 
 	@Override
 	public InfoDisplayContributorFieldType getType() {
-		return InfoDisplayContributorFieldType.TEXT;
+		return InfoDisplayContributorFieldType.URL;
 	}
 
 	@Override
 	public String getValue(FileEntry fileEntry, Locale locale) {
-		return fileEntry.getVersion();
+		try {
+			return _dlURLHelper.getDownloadURL(
+				fileEntry, fileEntry.getFileVersion(), null, StringPool.BLANK);
+		}
+		catch (Exception pe) {
+			return null;
+		}
 	}
+
+	@Reference
+	private DLURLHelper _dlURLHelper;
 
 }
