+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/info/display/contributor/field/FileEntryFileNameInfoDisplayContributorField.java
@@ -16,40 +16,33 @@ package com.liferay.document.library.web.internal.info.display.contributor.field
 
 import com.liferay.info.display.contributor.field.InfoDisplayContributorField;
 import com.liferay.info.display.contributor.field.InfoDisplayContributorFieldType;
-import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.repository.model.FileEntry;
-import com.liferay.portal.kernel.util.FastDateFormatFactoryUtil;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
-import java.text.Format;
-
 import java.util.Locale;
-import java.util.ResourceBundle;
 
 import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Eudaldo Alonso
+ * @author Alejandro Tardín
  */
 @Component(
 	property = "model.class.name=com.liferay.document.library.kernel.model.DLFileEntry",
 	service = InfoDisplayContributorField.class
 )
-public class DLFileEntryPublishDateInfoDisplayContributorField
+public class FileEntryFileNameInfoDisplayContributorField
 	implements InfoDisplayContributorField<FileEntry> {
 
 	@Override
 	public String getKey() {
-		return "publishDate";
+		return "fileName";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
-			locale, getClass());
-
-		return LanguageUtil.get(resourceBundle, "publish-date");
+		return LanguageUtil.get(
+			ResourceBundleUtil.getBundle(locale, getClass()), "file-name");
 	}
 
 	@Override
@@ -59,14 +52,7 @@ public class DLFileEntryPublishDateInfoDisplayContributorField
 
 	@Override
 	public String getValue(FileEntry fileEntry, Locale locale) {
-		if (fileEntry.getModifiedDate() == null) {
-			return StringPool.BLANK;
-		}
-
-		Format dateFormatDateTime = FastDateFormatFactoryUtil.getDateTime(
-			locale);
-
-		return dateFormatDateTime.format(fileEntry.getModifiedDate());
+		return fileEntry.getFileName();
 	}
 
 }
