+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/info/display/contributor/field/FileEntryMimeTypeInfoDisplayContributorField.java
@@ -19,7 +19,6 @@ import com.liferay.info.display.contributor.field.InfoDisplayContributorFieldTyp
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.TextFormatter;
 
 import java.util.Locale;
 
@@ -32,18 +31,18 @@ import org.osgi.service.component.annotations.Component;
 	property = "model.class.name=com.liferay.document.library.kernel.model.DLFileEntry",
 	service = InfoDisplayContributorField.class
 )
-public class DLFileEntrySizeInfoDisplayContributorField
+public class FileEntryMimeTypeInfoDisplayContributorField
 	implements InfoDisplayContributorField<FileEntry> {
 
 	@Override
 	public String getKey() {
-		return "size";
+		return "mimeType";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
 		return LanguageUtil.get(
-			ResourceBundleUtil.getBundle(locale, getClass()), "size");
+			ResourceBundleUtil.getBundle(locale, getClass()), "content-type");
 	}
 
 	@Override
@@ -53,7 +52,7 @@ public class DLFileEntrySizeInfoDisplayContributorField
 
 	@Override
 	public String getValue(FileEntry fileEntry, Locale locale) {
-		return TextFormatter.formatStorageSize(fileEntry.getSize(), locale);
+		return fileEntry.getMimeType();
 	}
 
 }
