+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/info/display/contributor/field/FileEntryPublishDateInfoDisplayContributorField.java
@@ -14,57 +14,59 @@
 
 package com.liferay.document.library.web.internal.info.display.contributor.field;
 
-import com.liferay.document.library.util.DLURLHelper;
 import com.liferay.info.display.contributor.field.InfoDisplayContributorField;
 import com.liferay.info.display.contributor.field.InfoDisplayContributorFieldType;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.repository.model.FileEntry;
+import com.liferay.portal.kernel.util.FastDateFormatFactoryUtil;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
+import java.text.Format;
+
 import java.util.Locale;
+import java.util.ResourceBundle;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Alejandro Tardín
+ * @author Eudaldo Alonso
  */
 @Component(
 	property = "model.class.name=com.liferay.document.library.kernel.model.DLFileEntry",
 	service = InfoDisplayContributorField.class
 )
-public class DLFileEntryDownloadURLInfoDisplayContributorField
+public class FileEntryPublishDateInfoDisplayContributorField
 	implements InfoDisplayContributorField<FileEntry> {
 
 	@Override
 	public String getKey() {
-		return "downloadURL";
+		return "publishDate";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(
-			ResourceBundleUtil.getBundle(locale, getClass()), "download-url");
+		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
+			locale, getClass());
+
+		return LanguageUtil.get(resourceBundle, "publish-date");
 	}
 
 	@Override
 	public InfoDisplayContributorFieldType getType() {
-		return InfoDisplayContributorFieldType.URL;
+		return InfoDisplayContributorFieldType.TEXT;
 	}
 
 	@Override
 	public String getValue(FileEntry fileEntry, Locale locale) {
-		try {
-			return _dlURLHelper.getDownloadURL(
-				fileEntry, fileEntry.getFileVersion(), null, StringPool.BLANK);
+		if (fileEntry.getModifiedDate() == null) {
+			return StringPool.BLANK;
 		}
-		catch (Exception pe) {
-			return null;
-		}
-	}
 
-	@Reference
-	private DLURLHelper _dlURLHelper;
+		Format dateFormatDateTime = FastDateFormatFactoryUtil.getDateTime(
+			locale);
+
+		return dateFormatDateTime.format(fileEntry.getModifiedDate());
+	}
 
 }
