+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/info/display/contributor/field/FileEntrySizeInfoDisplayContributorField.java
@@ -19,6 +19,7 @@ import com.liferay.info.display.contributor.field.InfoDisplayContributorFieldTyp
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
+import com.liferay.portal.kernel.util.TextFormatter;
 
 import java.util.Locale;
 
@@ -31,18 +32,18 @@ import org.osgi.service.component.annotations.Component;
 	property = "model.class.name=com.liferay.document.library.kernel.model.DLFileEntry",
 	service = InfoDisplayContributorField.class
 )
-public class DLFileEntryMimeTypeInfoDisplayContributorField
+public class FileEntrySizeInfoDisplayContributorField
 	implements InfoDisplayContributorField<FileEntry> {
 
 	@Override
 	public String getKey() {
-		return "mimeType";
+		return "size";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
 		return LanguageUtil.get(
-			ResourceBundleUtil.getBundle(locale, getClass()), "content-type");
+			ResourceBundleUtil.getBundle(locale, getClass()), "size");
 	}
 
 	@Override
@@ -52,7 +53,7 @@ public class DLFileEntryMimeTypeInfoDisplayContributorField
 
 	@Override
 	public String getValue(FileEntry fileEntry, Locale locale) {
-		return fileEntry.getMimeType();
+		return TextFormatter.formatStorageSize(fileEntry.getSize(), locale);
 	}
 
 }
